refactor(store_creator): structural and naming changes to store creator count
