+++ crates/core/src/db/queries/nft_count.rs
@@ -7,7 +7,7 @@ use diesel::{
     query_builder::QueryFragment,
     query_source::joins::{Inner, Join, JoinOn},
     serialize::ToSql,
-    sql_types::Text,
+    sql_types::{Array, Text},
     AppearsOnTable,
 };
 
@@ -17,6 +17,7 @@ use crate::{
         tables::{
             bid_receipts, current_metadata_owners, listing_receipts, metadata_creators, metadatas,
         },
+        models::StoreCreatorCount,
         Connection,
     },
     error::prelude::*,
@@ -218,4 +219,31 @@ where
         .count()
         .get_result(conn)
         .context("failed to load listed nfts count")
+}
+
+const STORE_CREATOR_QUERY: &str = r"
+select
+    sc.creator_address as store_creator,
+    count(distinct mc.metadata_address)::bigint as nfts
+
+from store_creators sc
+    inner join metadata_creators mc
+        on (mc.creator_address = sc.creator_address)
+
+where sc.creator_address = any($1) AND mc.verified
+group by sc.creator_address;
+ -- $1: store creator addresses::text[]";
+
+/// Count the number of nfts created by a creator
+///
+/// # Errors
+/// This function fails if the underlying SQL query returns an error
+pub fn store_creators(
+    conn: &Connection,
+    store_creators: impl ToSql<Array<Text>, Pg>,
+) -> Result<Vec<StoreCreatorCount>> {
+    diesel::sql_query(STORE_CREATOR_QUERY)
+        .bind(store_creators)
+        .load(conn)
+        .context("Failed to load store creators counts")
 }
