+++ crates/graphql/src/schema/dataloaders/store_creator.rs
@@ -1,9 +1,10 @@
-use indexer_core::db::tables::twitter_handle_name_services;
+use indexer_core::db::{queries::nft_count, tables::twitter_handle_name_services};
 use objects::store_creator::StoreCreator;
 use scalars::{markers::StoreConfig, PublicKey};
 use tables::store_creators;
 
 use super::prelude::*;
+use crate::schema::objects::store_creator::StoreCreatorCount;
 
 #[async_trait]
 impl TryBatchFn<PublicKey<StoreConfig>, Vec<StoreCreator>> for Batcher {
@@ -28,6 +29,22 @@ impl TryBatchFn<PublicKey<StoreConfig>, Vec<StoreCreator>> for Batcher {
         Ok(rows
             .into_iter()
             .map(|c| (c.1.store_config_address.clone(), c.try_into()))
+            .batch(addresses))
+    }
+}
+
+#[async_trait]
+impl TryBatchFn<PublicKey<StoreCreator>, Option<StoreCreatorCount>> for Batcher {
+    async fn load(
+        &mut self,
+        addresses: &[PublicKey<StoreCreator>],
+    ) -> TryBatchMap<PublicKey<StoreCreator>, Option<StoreCreatorCount>> {
+        let db = self.db()?;
+        let rows = nft_count::store_creators(&db, addresses)?;
+
+        Ok(rows
+            .into_iter()
+            .map(|s| (s.store_creator.clone(), s.try_into()))
             .batch(addresses))
     }
 }
